fix bugs

+++ src/main/java/com/gp/vaadin/demo/hotel/HotelUI.java
@@ -39,7 +39,7 @@ public class HotelUI extends UI {
 	final TextField filterByName = new TextField();
 	final TextField filterByAddress = new TextField();
 	final Button addHotel = new Button("Add hotel");
-	final Button deleteHotel = new Button("Delete hotel");
+	final static Button deleteHotel = new Button("Delete hotel");
 	private HotelEditForm form = new HotelEditForm(this);
 	final Hotel hotel = new Hotel();
 	
@@ -54,35 +54,50 @@ public class HotelUI extends UI {
     	
     	deleteHotel.setEnabled(false);
     	deleteHotel.addClickListener(e ->{
+     	if	(hotelGrid.getSelectedItems().iterator().hasNext()) {
     		Hotel delCandidate = hotelGrid.getSelectedItems().iterator().next();
     		hotelService.delete(delCandidate);
     		deleteHotel.setEnabled(false);
     		updateList();
     		form.setVisible(false);
-    	});
+    	 	}
+    	 	else {
+        		deleteHotel.setEnabled(false);
+        		updateList();
+    		 	form.setHotel(null);
+    		 	form.setVisible(false);
+    		 	deleteHotel.setEnabled(false);}});
     	
-        hotelGrid.setItems(hotelService.findAll());
-        hotelGrid.removeColumn("url"); 
-       
         Column<Hotel,?> htmlColumn = hotelGrid.addColumn(e ->
         "<a href='" + e.getUrl() + "' target='_top'>"+e.getUrl()+"</a>",
         new HtmlRenderer()).setCaption("Url");
+        
+        hotelGrid.setItems(hotelService.findAll());
+        hotelGrid.removeColumn("url"); 
     	hotelGrid.setColumnOrder("name", "address", "rating", "category");
-    	//hotelGrid.setWidth(100, Unit.PERCENTAGE);
     	hotelGrid.asSingleSelect().addValueChangeListener(e -> {
     		if(e.getValue() != null) {
     			deleteHotel.setEnabled(true);
+    			form.setEnabled(true);
     			form.setHotel(e.getValue());}});
     	 
-    	filterByName.addValueChangeListener(e -> updateList());
+    	filterByName.addValueChangeListener(e ->{updateList();
+    	if	(!hotelGrid.getSelectedItems().iterator().hasNext()) {
+    		form.setVisible(false);
+    		deleteHotel.setEnabled(false);}});
     	filterByName.setValueChangeMode(ValueChangeMode.LAZY);
     	filterByName.setPlaceholder("find by name");
     	
-    	filterByAddress.addValueChangeListener(e -> updateList());
+    	filterByAddress.addValueChangeListener(e ->{updateList();
+    	if	(!hotelGrid.getSelectedItems().iterator().hasNext()) {
+    		form.setVisible(false);
+    		deleteHotel.setEnabled(false);}});
     	filterByAddress.setValueChangeMode(ValueChangeMode.LAZY);
     	filterByAddress.setPlaceholder("find by address");
     	
-    	addHotel.addClickListener(e -> form.setHotel(new Hotel()));
+    	addHotel.addClickListener(e -> {
+    		form.setEnabled(true);
+    		form.setHotel(new Hotel());});
     	
         layout.addComponents(controls, content);
         setContent(layout);
